Wrapped nav container in banner section
